refactored supervised output connector and prepared for unsupervised output connector
